高さを揃えて文字数制限する
